requirements.txt
# Core ML/vision
opencv-python
numpy==1.26.4
pandas
scipy
Pillow
matplotlib
seaborn
scikit-learn

# PyTorch & ecosystem
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.6.0
torchvision==0.21.0
torchtext
fvcore
einops
timm

# Diffusion, transformers, and related
transformers==4.43.2  
diffusers==0.26.0
accelerate==1.6.0
xformers==0.0.29.post2 --find-links https://download.pytorch.org/whl/cu118
triton==3.2.0
peft==0.15.0

# 360/geometry
py360convert==1.0.3

# Super-resolution, GAN
realesrgan==0.3.0
basicsr

# Language & prompt
open_clip_torch

# Progress, utils
joblib
tqdm
gc

# Misc
plyfile